refactor(runs): flatten backends/ into backend.py and drop leftover scaffolding
The backends/ package was shaped for two implementations; after the
offline-mode cut it held one. Backend (protocol), EvalsBackend and the
disabled no-op now share backend.py, next to the contract they satisfy.

Also removed: @runtime_checkable on both protocols (no isinstance check
anywhere), the normalize_base_url helper (inlined into PlatformClient —
the prime_traces equivalent is private), and the top-level
build_samples / trace_to_sample aliases (prime_runs.projection is the
spelling verifiers already uses). The pyproject dependency comment no
longer claims a hand-rolled retry loop.

No behaviour change; 161 tests unchanged.
