+++ packages/prime-runs/pyproject.toml
@@ -14,10 +14,10 @@ authors = [
 # consumer of this SDK. Nothing here may pull in prime, verifiers, typer,
 # rich or textual, directly or transitively.
 #
-# Kept to what is actually imported. The other SDKs carry `pydantic` (they model
-# response bodies) and `tenacity` (they retry through it); this package models
-# nothing and hand-rolls its retry loop in `_http.py`, so neither belongs in the
-# dependency tree of a package that lands inside verifiers and prime-rl.
+# Kept to what is actually imported. This package models no response bodies
+# (so no direct `pydantic`; it arrives transitively through prime-traces) and
+# takes its error mapping and backoff from `prime_traces.core` rather than
+# `tenacity`. Nothing is added to what verifiers and prime-rl already carry.
 dependencies = [
     "httpx>=0.25.0",
     "prime-traces>=0.0.2",
